macro_tools : cleaning
