+++ module/core/iter_tools/src/iter.rs
@@ -61,6 +61,29 @@ pub( crate ) mod private
 
 }
 
+#[ doc( inline ) ]
+#[ allow( unused_imports ) ]
+#[ cfg( feature = "enabled" ) ]
+pub use protected::*;
+
+/// Protected namespace of the module.
+#[ cfg( feature = "enabled" ) ]
+pub mod protected
+{
+  #[ doc( inline ) ]
+  #[ allow( unused_imports ) ]
+  pub use super::orphan::*;
+}
+
+/// Orphan namespace of the module.
+#[ cfg( feature = "enabled" ) ]
+pub mod orphan
+{
+  #[ doc( inline ) ]
+  #[ allow( unused_imports ) ]
+  pub use super::exposed::*;
+}
+
 /// Exposed namespace of the module.
 pub mod exposed
 {
@@ -117,7 +140,6 @@ pub mod exposed
   pub use std::iter::zip;
 
 }
-
 
 /// Prelude to use essentials: `use my_module::prelude::*`.
 pub mod prelude
