+++ module/core/iter_tools/tests/inc/basic_test.rs
@@ -1,31 +1,21 @@
+#[ allow( unused_imports ) ]
 use super::*;
+#[ allow( unused_imports ) ]
 use TheModule::*;
 
 //
 
-tests_impls!
+#[ test ]
+#[ cfg( feature = "enabled" ) ]
+fn basic()
 {
-
-  #[ test ]
-  #[ cfg( feature = "enabled" ) ]
-  fn basic()
-  {
-    // test.case( "basic" );
-    let src = vec![ 1, 2, 3 ];
-    let exp = ( vec![ 2, 3, 4 ], vec![ 0, 1, 2 ] );
-    let got : ( Vec< _ >, Vec< _ > ) = src.iter().map( | e |
-    {(
-      e + 1,
-      e - 1,
-    )}).multiunzip();
-    a_id!( got, exp );
-  }
-
-}
-
-//
-
-tests_index!
-{
-  basic,
+  // test.case( "basic" );
+  let src = vec![ 1, 2, 3 ];
+  let exp = ( vec![ 2, 3, 4 ], vec![ 0, 1, 2 ] );
+  let got : ( Vec< _ >, Vec< _ > ) = src.iter().map( | e |
+  {(
+    e + 1,
+    e - 1,
+  )}).multiunzip();
+  a_id!( got, exp );
 }
